opencrew/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
passlib[bcrypt]
pydantic[email]
pydantic-settings
sqlalchemy
psycopg2-binary
alembic
celery[redis]
python-jose[cryptography]
httpx # Added for async requests in Auth0 validation
requests
beautifulsoup4
playwright
qdrant-client
sentence-transformers
numpy
stripe
PyPDF2
